Reogranize names/paths
