+++ src/Messages/Models/NodeInfoMessage.js
