+++ tests/MessageEncoder.js
@@ -1,17 +1,17 @@
 import test from 'ava'
-import MessageEncoder from '../src/MessageEncoder.js'
-import CloseMessage from '../src/Messages/CloseMessage.js'
-import PingMessage from '../src/Messages/PingMessage.js'
-import ResponseMessage from '../src/Messages/ResponseMessage.js'
-import FragmentMessage from '../src/Messages/FragmentMessage.js'
-import NodeInfoMessage from '../src/Messages/NodeInfoMessage.js'
-import GetGenerationMessage from '../src/Messages/GetGenerationMessage.js'
+import MessageEncoder from '../src/Messages/Encoder/MessageEncoder.js'
+import CloseMessage from '../src/Messages/Models/CloseMessage.js'
+import PingMessage from '../src/Messages/Models/PingMessage.js'
+import ResponseMessage from '../src/Messages/Models/ResponseMessage.js'
+import FragmentMessage from '../src/Messages/Models/FragmentMessage.js'
+import NodeInfoMessage from '../src/Messages/Models/NodeInfoMessage.js'
+import GetGenerationMessage from '../src/Messages/Models/GetGenerationMessage.js'
 import Peer from '../src/Peer.js'
 import ChainObject from '../src/ChainObjects/ChainObject.js'
-import TransactionsMessage from '../src/Messages/TransactionsMessage.js'
-import MicroBlockMessage from '../src/Messages/MicroBlockMessage.js'
-import KeyBlockMessage from '../src/Messages/KeyBlockMessage.js'
-import GenerationMessage from '../src/Messages/GenerationMessage.js'
+import TransactionsMessage from '../src/Messages/Models/TransactionsMessage.js'
+import MicroBlockMessage from '../src/Messages/Models/MicroBlockMessage.js'
+import KeyBlockMessage from '../src/Messages/Models/KeyBlockMessage.js'
+import GenerationMessage from '../src/Messages/Models/GenerationMessage.js'
 
 const encoder = new MessageEncoder()
